Handle not enough points for mean_curv calc
I handled it by setting c_atrack::mean_curv to float('nan')
if n_curv == 0.  This results in "nan" being saved to the
files for those tracks.

In the test ipro 0111v05b au 0112c, v021 is a track for which
there is only one track and it doesn't have enough points for
a curv calculation.